requirements.txt
affine==2.4.0
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
billiard==4.2.0
cachetools==5.3.2
celery==5.3.6
celery-progress==0.3
certifi==2023.7.22
cftime==1.6.3
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cligj==0.7.2
dataex-client @ https://github.com/nzahasan/dataex-client/zipball/master#sha256=2d32527073997839d2d8760f00c85d50b53d87e82cc3a9caa7cde4eb6401b098
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.5
django-celery-results==2.5.1
django-cors-headers==4.2.0
django-filter==23.2
django-import-export==3.3.7
django-pagination==1.0.7
django-pandas==0.6.6
django-redis==5.4.0
django-simple-pagination==1.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
earthengine-api==0.1.376
et-xmlfile==1.1.0
fiona==1.9.5
geopandas==0.14.0
google-api-core==2.12.0
google-api-python-client==2.105.0
google-auth==2.23.3
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.2.0
google-cloud-core==2.3.3
google-cloud-storage==2.12.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
greenlet==3.0.3
gunicorn==21.2.0
httplib2==0.22.0
idna==3.4
importlib-metadata==6.8.0
kombu==5.3.6
MarkupPy==1.14
netCDF4==1.6.5
numpy==1.25.2
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.2
panda==0.3.1
pandas==2.1.0
prompt-toolkit==3.0.43
protobuf==4.24.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyJWT==2.8.0
PyMySQL==1.1.0
pyparsing==3.1.1
pyproj==3.6.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
rasterio==1.3.9
redis==5.0.3
requests==2.31.0
requests-oauthlib==2.0.0
rioxarray==0.15.0
rsa==4.9
shapely==2.0.2
six==1.16.0
snuggs==1.4.7
SQLAlchemy==2.0.20
sqlparse==0.4.4
tablib==3.5.0
tabulate==0.9.0
termcolor==2.4.0
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.6.0
xarray==2024.3.0
xlrd==2.0.1
xlwt==1.3.0
yaspin==3.0.1
zipp==3.17.0
